+++ R/GenotypeTableFunctions.R
@@ -57,44 +57,63 @@ readGenotypeTableFromPath <- function(path, keepDepth = FALSE, sortPositions = F
 #'
 #' @param tasObj An object of class \code{TasselGenotypePenotype}.
 #' @param useRef Use reference alleles or major alleles at sites. If
-#'    \code{FALSE}, major alleles will be used.
+#'    \code{FALSE}, major alleles will be used. Defaults to \code{FALSE}.
+#' @param coerceDosageToInt Should dosage array be returned as \code{integer}
+#'    values? If \code{FALSE}, dosage array will be returned as type
+#'    \code{raw} byte values. Returning \code{raw} byte values. Will greatly
+#'    save on memory. Defaults to \code{TRUE}.
+#' @param verbose Should messages be displayed to console? Defaults to
+#'    \code{FALSE}.
 #'
 #' @return Returns a \code{SummarizedExperiment} of TASSEL genotype data.
 #'
 #' @importFrom rJava .jcall
 #' @importFrom rJava is.jnull
+#' @importFrom rJava .jevalArray
 #' @importFrom SummarizedExperiment SummarizedExperiment
 #' @export
-getSumExpFromGenotypeTable <- function(tasObj, useRef = FALSE) {
-    jGT <- getGenotypeTable(tasObj)
+getSumExpFromGenotypeTable <- function(tasObj,
+                                       useRef = FALSE,
+                                       coerceDosageToInt = TRUE,
+                                       verbose = FALSE) {
 
     if (class(tasObj) != "TasselGenotypePhenotype") {
         stop("`tasObj` must be of class `TasselGenotypePhenotype`")
     }
 
+    jGT <- getGenotypeTable(tasObj)
+
     if (rJava::is.jnull(jGT)) {
-        stop("TASSEL genotype table not detected.")
+        stop("TASSEL genotype object not found")
     }
 
+    # Create SumExp components (DF and ranges)
     sampleDF <- sampleDataFrame(jGT)
     genomicRangesDF <- genomicRanges(jGT)
 
+    # Create and return byte array from TASSEL
+    if (verbose) message("Generating byte array...")
     jc <- rJava::J("net/maizegenetics/plugindef/GenerateRCode")
-
-    genoCallIntArray <- jc$genotypeTableToDosageIntArray(
+    genoCallByteArray <- jc$genotypeTableToDosageByteArray(
         jGT,
         useRef
     )
+    if (verbose) message("Returning Java byte array to R...")
+    dosMat <- lapply(genoCallByteArray, rJava::.jevalArray)
+    if (coerceDosageToInt) {
+        if (verbose) message("Coercing to integer...")
+        dosMat <- lapply(dosMat, as.integer)
+    }
+    if (verbose) message("Transforming to SummarizedExperiment...")
+    dosMat <- simplify2array(dosMat)
 
-    SummarizedExperiment::SummarizedExperiment(
-        assays = matrix(
-            genoCallIntArray,
-            length(genomicRangesDF),
-            byrow = TRUE
-        ),
+    se <- SummarizedExperiment::SummarizedExperiment(
+        assays = dosMat,
         rowRanges = genomicRangesDF,
         colData = sampleDF
     )
+    if (verbose) message("Finished.")
+    return(se)
 }
 
 
@@ -112,3 +131,55 @@ getGenotypeTable <- function(jtsObject) {
         return(rJava::.jnull())
     }
 }
+
+
+## Return min/max physical positions from genotype tables (house keeping)
+getMinMaxPhysPositions <- function(tasObj) {
+    if (class(tasObj) != "TasselGenotypePhenotype") {
+        stop("`tasObj` must be of class `TasselGenotypePhenotype`")
+    }
+
+    jGenoTable <- getGenotypeTable(tasObj)
+    if (rJava::is.jnull(jGenoTable)) {
+        stop("TASSEL genotype object not found")
+    }
+
+    javaGT <- getGenotypeTable(tasObj)
+
+    positions <- javaGT$positions()
+    chroms <- rJava::.jevalArray(javaGT$chromosomes())
+    posLS  <- lapply(chroms, javaGT$firstLastSiteOfChromosome)
+
+    physPos <- lapply(posLS, function(x) {
+        firstPos <- positions$get(x[1])
+        lastPos  <- positions$get(x[2])
+        return(c(firstPos$getPosition(), lastPos$getPosition()))
+    })
+    names(physPos) <- sapply(chroms, function(x) x$getName())
+    return(physPos)
+}
+
+
+## Return min/max physical positions from genotype tables (house keeping)
+getMinMaxVarSites <- function(tasObj) {
+    if (class(tasObj) != "TasselGenotypePhenotype") {
+        stop("`tasObj` must be of class `TasselGenotypePhenotype`")
+    }
+
+    jGenoTable <- getGenotypeTable(tasObj)
+    if (rJava::is.jnull(jGenoTable)) {
+        stop("TASSEL genotype object not found")
+    }
+
+    javaGT <- getGenotypeTable(tasObj)
+
+    positions <- javaGT$positions()
+    chroms <- rJava::.jevalArray(javaGT$chromosomes())
+    posLS  <- lapply(chroms, javaGT$firstLastSiteOfChromosome)
+
+    names(posLS) <- sapply(chroms, function(x) x$getName())
+    return(posLS)
+}
+
+
+